Remove the namespace declaration
After installing the package via pip, the __init__.py from the source
distribution does not exist. It is instead replaced by:

  pyga-2.4.2-py2.7-nspkg.pth

Which does some magic to make sure that sys.modules is updated correctly

I am installing pyga and deploying it to appspot and the import fails
because __init__.py is missing.